Implement bare minimum builtin `math`
This is the second baby step in resolving #3157. Implement a bare minimum
builtin `math` command. This doesn't even return the results so that fish
scripts can use them. It simply writes the result directly to stdout. This
is solely to ensure that fish can be built and run in the Travis build
environments. This is okay since anyone running `builtin math` today is
already getting an error response.

Also, much more work is needed to support bare var references, multiple
result values, etc.