Implement `history count` and `history item`
The `history` variable is incredibly inefficient. Just doing `set -g`
might take an extra 200 ms if you have 30,000 entries in your history.
It's also very limited since it only provides the command line but not
when it was run or the paths that fish extracted from the command.

So replace the `history` var with a `history item` command.

Fixes #4366