Split command substitutions on NUL if one is found
Currently, there is no easy way to use e.g. `find -print0` in command substitutions.

What this does is check if a NUL is in the comsub output and then split on that instead of newline,
which makes `something (find -print0)` just work.

The one fly in the ointment is if there is a command that uses NULs to separate its output,
but doesn't print a trailing NUL if there is just one "thing" to print.
That would cause us to fall back to splitting on newlines, which might do the wrong thing.

So far, everything I've tested (`find`, `git config -z`) seemed to print a trailing NUL.

This implements a proposal by @cben.

Fixes #3164.