[WIP] Use ncurses for handling window size
This commit drops direct handling of window size, including the
TIOCGWINSZ ioctl and reading/writing of the LINES and COLUMNS
environment variables, all of which are already performed by (n)curses.

As we already have a dependency on ncurses, it doesn't make sense to
make use of its APIs and reduce the NIH code in fish, relying on a
library developed and maintained by people most concerned with terminals
is probably the right way to go.

* Lightly tested only at this time.
* Most old code was commented out and not removed.
* Build system still checks for TIOCGWINSZ.
* SIGWINCH handlers left in place since not all versions of (n)curses
  handle SIGWINCH and supply a KEY_RESIZE input when the terminal is
  resized.
* Performance might be improved by no longer obtaining a mutex to obtain
  the screen size since a single atomic read/write is performed to the
  invalidation volatile bool variable as ncurses handles state
  internally, presuming that the `getmaxyx()` ncurses call is
  thread-safe.