Don't execute completions automatically if head does not exist
This fixes the issues with fish automatically spewing error output to
the console caused by autoloaded completions attempting to generate
results to provide the autosuggestion thread with data.

What we do is prevent completions from being considered as an
autosuggestion source if the command being typed out is not a
fully-formed head that exists. This only addresses *automatically*
generated errors, such as that when typing in `zfs` or `git` at the
prompt on an installation that does not have these utilities. This does
_not_ prevent errors caused by typing in `zfs <TAB>` or `git <TAB>` when
these tools are not installed.

Closes #2365 and possibly others; it has many duplicates such as #4688.