Use `make` to call whatever build system is available
* `ninja` via cmake, if both are installed
* `make` via cmake, if cmake is installed
* GNU `make`, if `./configure` is present (i.e. tarball and not git)
* `autoconf` + gnu `make` if `autoconf` is installed and `./configure.ac`
  is present,
* or error out, otherwise

This works because GNU Make will use `GNUmakefile`, if present, over any
other `Makefile` in the same directory. We call autoreconf (if needed)
then `./configure` (passing in PREFIX as --prefix) to create the
`GNUmakefile` in the current directory, then "recursively" (but not
really) call `make` once more against the CWD. This time, GNU `make`
will see and use the `GNUmakefile` insead.

When we "recursively" call `make` a second time, we pass along all
options that we were given at the command line, making it seem like the
user had called `make` directly.