Cache function descriptions to display when not yet autoloaded
Addresses long-standing bug #328 (which dates back to fish 1.0) by
caching the description for auto-loaded functions when they are first
added in a thread-safe, lightweight, and performant manner (thanks to
lmdb).

While function descriptions are cached, they will not be candidates for
completion unless the function can be autoloaded. The last function
autloaded always replaces previously cached descriptions (i.e. in the
event of multiple scopes).

Closes #328.