Normalize tokenization of non-standard `[` usages
This enables the tokenizer to handle weird cases like those demonstrated
in this comment [0] on #2485. We basically expand support for `[` as
plain text to cover more cases than just the `[` alias for `test`.

Honestly, `[` and `]` should generally be quoted as they have very
special meaning to the shell, but if someone wants to do something
creative with them (@cben, still interested?) this will at least
standardize their behavior and remove ambiguous parsing cases.

[0]: https://github.com/fish-shell/fish-shell/issues/2485#issuecomment-167235194