Revert "Disable ONLCR mapping of NL output to CR-NL"
This reverts commit 3f820f0edf637bc5769602dd7fff1fb22b2329cc.
While the premise described by @nbuwe is sound in #4505, we are now
apparently relying on this behavior is some places (although
inadvertently as there doesn't seem to be a deliberate acknowledgement
of that anywhere).

Turning off ONLCR causes things like indented multiline commands to not
appear correct at the tty (subsequent lines appear both at column 0 and
again indented).